requirements.txt
django>=2
channels>=2
asgi_redis
channels_redis
psycopg2
django-htmlmin
django-ipware
requests
daphne>=2
django-redis
beautifulsoup4
lxml
pyyaml